Switch to i18n
More idiomatic to use internationalisation over localisation. I think.

Also relaxed the langauge switch on 'en', as it seems to be
generating 'en_US' rather than 'en' when GitHub generates it. To be
looked into in future, but it currently just doesn't work at all
(except on my machine, lol).